"ads2_m8_a1_1: 2/6 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 1/1 "

+++ practice/M8/1. Shortest Paths/DijkstraSP.java
@@ -108,8 +108,8 @@ class DijkstraSP {
      *
      * @return shortest distance between two vertices.
      */
-    public int getDistance(final int vertex) {
-        int sum = 0;
+    public double getDistance(final int vertex) {
+        double sum = 0;
         for (Edge each : pathTo(vertex)) {
             sum += each.weight();
         }
